Update on "all_gather_bucketing fx pass"
Porting passes to bucket all_gathers

The main logic of the pass is done via 
1. Searching for all all_gathers from the buckets
2. Inserting the trace of the function instead of all_gathers:
```
def all_gather_merge_fn_to_trace(  # type: ignore[no-untyped-def]
    ag_inputs, group_size: int, group_name: str, dtype: torch.dtype
) -> list[torch.Tensor]:
    inp_sizes = [ag_in.shape for ag_in in ag_inputs]
    inp_split_sizes = [ag_in.numel() for ag_in in ag_inputs]
    ag_input_numel = sum(inp_split_sizes)
    device = ag_inputs[0].device
    rank = device.index
    new_ag_out = torch.empty(ag_input_numel * group_size, dtype=dtype, device=device)
    foreach_copy_dsts = torch.split(new_ag_out, inp_split_sizes)
    ag_inputs_reshaped = [ag_in.reshape(-1) for ag_in in ag_inputs]
    torch._foreach_copy_(foreach_copy_dsts, ag_inputs_reshaped)
    new_ag_in = new_ag_out.narrow(0, ag_input_numel * rank, ag_input_numel)
    wait_tensor = torch.ops.c10d_functional.wait_tensor(
        torch.ops._c10d_functional.all_gather_into_tensor_out.default(
            new_ag_in, group_size, group_name, out=new_ag_out
        )
    )
    new_ag_out_reshaped = wait_tensor.reshape(group_size, -1)
    outs = torch.split_with_sizes(
        new_ag_out_reshaped,
        inp_split_sizes,
        dim=1,
    )
    outs_reshaped = [out.reshape(shape) for out, shape in zip(outs, inp_sizes)]
    return outs_reshaped
```
Having special function that we trace allows to simplify the experimentation on changing what we are writing.
3. This graph will be inserted as early as possible - just after we found the last input to bucketing all_gathers

Copying tests from wconstab PR to test compatibility with reordering.
Test checks only compatibility, as because of (3) the joint all_gather will be scheduled already as early as possible and no space for reordering.

Will try to  come up to test actual reordering.


Pass changes:
Using mutation ops to match performance of fsdp, in future the perfect scenario will be to have only functional graph, that inductor does all memory optimizations on its own without mutable ops.


Inductor changes:
Adding foreach_copy_ lowering


cc H-Huang awgu wanchaol fegin fduwjj wz337 wconstab d4l3k voznesenskym penguinwu EikanWang jgong5 Guobing-Chen XiaobingSuper zhuhaozhe blzheng wenzhe-nrv jiayisunx ipiszy chenyang78 kadeng muchulee8 amjames chauhang aakhundov

[ghstack-poisoned]
